move components to seperate file into components folder
