+++ 01-starting-project/src/components/CoreConcept.jsx
@@ -1,0 +1,9 @@
+export default function CoreConcept({ image, title, description }) {
+  return (
+    <li>
+      <img src={image} alt="..." />
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </li>
+  );
+}
